+++ client/src/App.js
@@ -19,6 +19,7 @@ class App extends Component {
           <Route path="/map_tool" component={MapToolPage}/>
           <Route path="/api_test" component={ApiTestPage}/>
           <Route path="/surveys" component={SurveyListPage}/>
+          <Route path="/survey/:surveyId/session/:sessionId/activity/:activityIdx" component={SurveyPage}/>
           <Route path="/survey/:surveyId/session/:sessionId" component={SurveyPage}/>
           <Route path="/survey/:surveyId" component={SurveyPage}/>
         </Switch>