+++ client/src/index.js
@@ -10,7 +10,7 @@ import { BrowserRouter } from 'react-router-dom'
 ReactDOM.render(
   <BrowserRouter>
     <App />
-  </BrowserRouter>, 
+  </BrowserRouter>,
   document.getElementById('root')
 );
 registerServiceWorker();