+++ client/src/pages/SurveyPage.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
-import { find } from 'lodash';
+import { find, has } from 'lodash';
+import { Redirect } from 'react-router-dom';
 
 import { getSurveyDefinitions, createSession } from '../services/api';
 import Intro from '../components/Intro';
@@ -7,13 +8,27 @@ import FormActivity from '../components/FormActivity';
 
 class HomePage extends Component {
   state = {
-    surveyDefinitions: []
+    surveyDefinitions: [],
+    activity: null,
+    redirect: false,
   }
-  
+
   componentDidMount() {
     getSurveyDefinitions()
-      .then(res => this.setState({ surveyDefinitions: res }))
+      .then(res => this.setState({
+        surveyDefinitions: res
+      }))
       .catch(err => console.log(err));
+  }
+
+  static getDerivedStateFromProps(props, state) {
+    if (has(props, 'location.pathname') && props.location.pathname === state.redirect) {
+      return {
+        ...state,
+        redirect: false
+      }
+    }
+    return state;
   }
 
   // Get the current survey based on URL params
@@ -29,20 +44,50 @@ class HomePage extends Component {
     return this.props.match.params.sessionId;
   }
 
+  getActivityIndex() {
+    return parseInt(this.props.match.params.activityIdx, 10);
+  }
+
   handleBeginClick = () => {
     createSession()
       .then((res) => {
-        window.location.replace(`/survey/${this.getSurveyId()}/session/${res.id}`);
+        window.location.replace(
+          `/survey/${this.getSurveyId()}/session/${res.id}/activity/0`
+        );
       })
       .catch(err => console.log(err));
   }
 
+  handleSubmit = () => {
+    // TODO submit answers
+
+    const nextActivityIdx = this.getActivityIndex() + 1;
+    if ( nextActivityIdx >= this.getSurvey().activities.length){
+      // TODO render the completion screen
+      console.log('Survey complete');
+    } else {
+      this.setState({
+        redirect: `/survey/${this.getSurveyId()}/session/${this.getSessionId()}/activity/${nextActivityIdx}`
+      })
+    }
+  }
+
   render() {
+    // We put redirect paths in the state when its time to move to the next page.
+    // If there is a redirect path in state, we render a rediect.
+    // This is a bit of an oddness of react router.
+    if (this.state.redirect) {
+      return <Redirect to={{
+        pathname: this.state.redirect,
+        state: { redirect: null }
+      }} push />
+    }
+
     const surveyDefinition = this.getSurvey();
     if (!surveyDefinition) {
       return null;
     }
-    
+
     if (!this.getSessionId()) {
       return (
         <Intro
@@ -52,12 +97,12 @@ class HomePage extends Component {
         />
       );
     }
-    console.log('Rendering survey: ', JSON.stringify(surveyDefinition));
-    
+
     return (
       <div>
         <FormActivity
-          activity={surveyDefinition.activities[0]}
+          activity={surveyDefinition.activities[this.getActivityIndex()]}
+          onSubmit={this.handleSubmit}
         />
       </div>
     );